+++ test_2025_04_17/Block.h
@@ -6,8 +6,8 @@ class Block
 {
 protected:
 	BlockInfo blockInfo;
-	float m_downSpeed; // 落ちる速さ
-	float m_lateralSpeed; // 横移動の速さ
+	double m_downSpeed; // 落ちる速さ
+	double m_lateralSpeed; // 横移動の速さ
 
 public:
 	// コンストラクタ
